Rename scanner
